backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
databricks-sql-connector==2.9.4
pydantic==2.5.2
